Use build tags for double signer implementation
