+++ Makefile
@@ -73,7 +73,10 @@ test-e2e: check-go
 .PHONY: test-e2e-polybft
 test-e2e-polybft: check-go
 	go build -o artifacts/polygon-edge .
-	env EDGE_BINARY=${PWD}/artifacts/polygon-edge E2E_TESTS=true E2E_LOGS=true BYZANTINE_BINARY=${PWD}/e2e-polybft/e2e/polygon-edge-byzantine \
+	go build -tags doubleSigner -o artifacts/polygon-edge-double-signer .
+	env EDGE_BINARY=${PWD}/artifacts/polygon-edge \
+	BYZANTINE_BINARY=${PWD}/artifacts/polygon-edge-double-signer \
+	E2E_TESTS=true E2E_LOGS=true \
 	go test -v -timeout=1h30m ./e2e-polybft/e2e/...
 
 .PHONY: test-property-polybft
