+++ consensus/polybft/custom_multicast_noop.go
@@ -1,0 +1,8 @@
+//go:build !doubleSigner
+
+package polybft
+
+import ibftProto "github.com/0xPolygon/go-ibft/messages/proto"
+
+func (p *Polybft) customMulticastHandler(msg *ibftProto.Message) {
+}
